+++ src/gwenbotv3/database/handlers/gwensub_handler.py
@@ -3,34 +3,50 @@ from sqlite3 import Cursor
 from gwenbotv3 import SingletonLogger
 from gwenbotv3.database import connect
 from gwenbotv3.database import UserContext
+from gwenbotv3.database.get_context import context
+from gwenbotv3.database.handlers.user_handler import UserHandler
 
 
 class GwenSubHandler:
     def __init__(self):
         self.logger = SingletonLogger().get_logger()
+        self.user_handler = UserHandler()
 
     @connect
-    def fetch_sub(self, cur: Cursor, context: UserContext) -> bool:
+    def fetch_sub(self, cur: Cursor, ctx: UserContext) -> bool:
+        if not ctx.user:
+            return False
+
         res = cur.execute(
             "SELECT * FROM Subs WHERE user=? AND server=?",
-            (context.user.id, context.server.id),
+            (ctx.user.id, ctx.server.id),
         ).fetchone()
 
         return True if res else False
 
     @connect
-    def add_sub(self, cur: Cursor, context: UserContext) -> None:
+    def add_sub(self, cur: Cursor, ctx: UserContext) -> None:
+        if not ctx.user:
+            self.user_handler.insert_user(ctx.ctx)
+            ctx = context(ctx.ctx)
+
+        if not ctx.user:
+            return
+
         cur.execute(
             "INSERT INTO Subs(user, server) VALUES(?,?) "
             "ON CONFLICT(user, server) DO NOTHING",
-            (context.user.id, context.server.id),
+            (ctx.user.id, ctx.server.id),
         )
 
     @connect
-    def remove_sub(self, cur: Cursor, context: UserContext) -> bool:
+    def remove_sub(self, cur: Cursor, ctx: UserContext) -> bool:
+        if not ctx.user:
+            return False
+
         cur.execute(
             "DELETE FROM Subs WHERE user=? AND server=?",
-            (context.user.id, context.server.id),
+            (ctx.user.id, ctx.server.id),
         )
 
         if cur.rowcount < 1:
@@ -39,29 +55,40 @@ class GwenSubHandler:
         return True
 
     @connect
-    def fetch_blacklist(self, cur: Cursor, context: UserContext) -> bool:
+    def fetch_blacklist(self, cur: Cursor, ctx: UserContext) -> bool:
+        if not ctx.user:
+            return False
+
         res = cur.execute(
             "SELECT * FROM Blacklist WHERE user=? AND server=?",
-            (context.user.id, context.server.id),
+            (ctx.user.id, ctx.server.id),
         ).fetchone()
 
         return True if res else False
 
     @connect
-    def add_blacklist(self, cur: Cursor, context: UserContext, by_owner: bool = False):
+    def add_blacklist(self, cur: Cursor, ctx: UserContext, by_owner: bool = False):
+        if not ctx.user:
+            self.user_handler.insert_user(ctx.ctx)
+            ctx = context(ctx.ctx)
+
+        if not ctx.user:
+            return
+
         cur.execute(
             "INSERT INTO Blacklist(user, server, by_owner) VALUES (?,?,?) "
             "ON CONFLICT(user, server, by_owner) DO NOTHING",
-            (context.user.id, context.server.id, by_owner),
+            (ctx.user.id, ctx.server.id, by_owner),
         )
 
     @connect
-    def remove_blacklist(
-        self, cur: Cursor, context: UserContext, by_owner: bool = False
-    ):
+    def remove_blacklist(self, cur: Cursor, ctx: UserContext, by_owner: bool = False):
+        if not ctx.user:
+            return False
+
         cur.execute(
             "DELETE FROM Blacklist WHERE user=? AND server=? AND by_owner=?",
-            (context.user.id, context.server.id, by_owner),
+            (ctx.user.id, ctx.server.id, by_owner),
         )
 
         if cur.rowcount < 1:
